+++ CHANGELOG.md
@@ -2,5 +2,6 @@
 - 
 ### Fixed:
 - Fixed offhand usage
+- Fixed bow damage
 ### Changed:
-- 
+- Changed reforge loader